+++ src/native/test/loop/RunTests.c
@@ -47,7 +47,7 @@ int main(void) {
     int retVal = 0;
     runFixture(&AceUnitLoopTestFixture);
     if (runnerData->testCaseFailureCount != 0) {
-        fprintf(stderr, "Test Cases: %d  Errors: %d\n", runnerData->testCaseCount, runnerData->testCaseFailureCount);
+        fprintf(stderr, "Test Cases: %d  Errors: %d\n", (int) runnerData->testCaseCount, (int) runnerData->testCaseFailureCount);
         retVal = 1;
     }
     if (loopCount1 != 5) {
